TESJH
Removed "Test_Sandbox_2".
Replaced "Test_Sandbox_3" with "ClientSocket_Example" connects to existing socket and writes data. (You can do a read on that socket as well, but example not implemented yet)
Replaced "Test_Sandbox_4" with "ServerSocket_Example" creates a new socket and listens for a connection and then accepts then reads the data from the socket. (You can do a write on the socket as well, but example not implemented yet.)

Added "Accept" and "ReadAll" methods to "TES.Integration.IPSocket".

Created new project "TES.Integration.IPSocket2" as a ".NET Framework Class Library" and copied the "Base" class from "TES.Integration.IPSocket". I created this because "TES.Integration.IPSocket" is a ".NET Standard Class Library"(which is different from what we normally use and doesn't seem to work in NAV).